Trait astar_bump::SearchProblem

pub trait SearchProblem<N, C, I: Iterator<(N, C)>> {
    fn start(&self) -> N;
    fn is_end(&self, &N) -> bool;
    fn heuristic(&self, &N) -> C;
    fn neighbors(&self, at: &N) -> I;

    fn estimate_length(&self) -> Option<uint> { ... }
}

A SearchProblem is a description of the problem that will be solved with A*. Implementing this trait will describe the problem well enough that it can be solved without any more information. N is the type of one of the search-states and C is the type of the cost to get from one state to another.

Required Methods

fn start(&self) -> N

A state representing the start of the search.

fn is_end(&self, &N) -> bool

Check to see if a state is the goal state.

fn heuristic(&self, &N) -> C

A function that estimates the cost to get from a node to the end. heuristic(end_state) should always be 0.

fn neighbors(&self, at: &N) -> I

A function returning the neighbors of a search state along with the cost to get to that state.

Provided Methods

fn estimate_length(&self) -> Option<uint>

This method is used if an estimated length of the path is available.